chore: Move `providers` folder into `navigation` folder
